Show the eval's output even if it's false.
The code q{print "0"; "WOMBLE"} when evaluated produces stdout that evaluates to false.  Previously the output of an eval was tested for perl's idea of truth, and if found false, the return value of the eval was dumped.  This change checks the length of the output, and always shows the output if there was some output.

See freenode #perl 20171111 1813Z to 1822Z.